==> BIg Data Final Project/0 - OldData and Steamdata split.ipynb <==
import pandas as pd
path = r'C:\Users\AhmaDGoly\Desktop\Arshad\TERM 2\Big Data\Final\Codes\AllMsgsTill7-10-2023.csv'
savepath =r'C:\Users\AhmaDGoly\Desktop\Arshad\TERM 2\Big Data\Final\Codes\\'
df = pd.read_csv(path , encoding='utf-8-sig')
first_part = df[:200]
second_part = df[200:]

first_part.to_csv(savepath+'StreamData.csv', index=False , encoding='utf-8-sig')
second_part.to_csv(savepath+'OldData.csv', index=False , encoding='utf-8-sig')
