==> BIg Data Final Project/5 - cassandra.ipynb <==
#age in code ro nafahmidin khodetoon ye code joda ba estefade az StreamData_preprocessed.csv benevisin. yani
#ye code benevisin ke StreamData_preprocessed.csv ro load kone va satr hasho yeki yeki pass bede be codetoon.
#input: 7-column-rows
#output: just save it to cassandra. and write an app to get a hashtag as input and show the articles about that hashtag in python (show the output in console).
import json,datetime
from kafka import KafkaConsumer
import pandas as pd

topicToReceive = 'cassandra'
kafkaPort = 19092

def current_time():
    current_datetime = datetime.datetime.now()
    return current_datetime.strftime("[%Y-%m-%d_%H:%M:%S]")

consumer = KafkaConsumer(topicToReceive, bootstrap_servers=[f'127.0.0.1:{kafkaPort}'])

x = 0
for message in consumer:
    x = x + 1
    print(f"{current_time()}: {x}: Received a news from Preprocessing_Thread. Processing...")
    json_row = message.value.decode('utf-8')
    row = json.loads(json_row)
    receivedRowDF = pd.DataFrame([row])
    #az inja be bad kodetoon ro be ezaye har satre jadidi ke omad piade konin.
